+++ src/components2/ProjectList/ReportListItem.js
@@ -1,23 +1,14 @@
 import React, { PureComponent } from 'react';
 import styled from 'styled-components';
-import slugify from 'slugify';
 
 import config from '~/config';
-import { numberFormat, getRVALength } from '~/utils/utils';
 import Label from '~/components/Label';
 import Button from '~/components/Button';
 import HeartIcon from '~/images/heart.svg';
-import DraftMarker from '~/images/planning-icons/konzept-marker.png';
-import PlanningMarker from '~/images/planning-icons/planung-marker.png';
-import ExecutionMarker from '~/images/planning-icons/bau-marker.png';
-import ReadyMarker from '~/images/planning-icons/fertig-marker.png';
+import BikestandsIcon from '~/images/reports/bikestands-icon.svg';
+import DefaultPhotoSrc from '~/images/reports/landing-christin-hume-595752-unsplash.jpg';
 
-const icons = {
-  draft: DraftMarker,
-  planning: PlanningMarker,
-  execution: ExecutionMarker,
-  ready: ReadyMarker
-};
+import { getReportStatusCaption } from '~/pages/Reports/apiservice';
 
 const ItemWrapper = styled.div`
   margin: 8px 0;
@@ -37,7 +28,7 @@ const ItemContent = styled.div`
 
 const ItemHeader = styled.div``;
 
-const ItemImage = styled.img`
+const ItemImage = styled(BikestandsIcon)`
   position: absolute;
   right: 16px;
   top: 16px;
@@ -48,14 +39,15 @@ const ItemTitle = styled.div`
   font-size: 14px;
   color: ${config.colors.darkgrey};
   font-weight: 600;
+  margin-bottom: 3px;
 `;
 
 const ItemSubTitle = styled.div`
   font-size: 14px;
   font-weight: 700;
-  font-family: '${config.titleFont}', serif;
   color: ${config.colors.darkbg};
-  margin: .5em auto;
+  margin-top: 1em;
+  margin-bottom: 1em;
 `;
 
 const ItemFooter = styled.div`
@@ -114,20 +106,7 @@ const MapButton = styled(Button)`
   }
 `;
 
-const ProjectLength = ({ length, side, id }) => {
-  const rvaLength = getRVALength({ length, side, id });
-  if (rvaLength == null) return null;
-  return (
-    <>
-      | {numberFormat(rvaLength / 1000, 1)} km
-      {side === 2 && (
-        <> (beidseitige Planung mit je {numberFormat(length / 1000, 1)} km)</>
-      )}
-    </>
-  );
-};
-
-class ProjectListItem extends PureComponent {
+class ReportListItem extends PureComponent {
   constructor(props) {
     super(props);
     this.state = {
@@ -136,10 +115,8 @@ class ProjectListItem extends PureComponent {
   }
 
   onClick = () => {
-    const { id, street_name: name } = this.props;
-    const slug = name ? slugify(name) : '';
-    const url = `${config.routes.projects}/${id}/${slug.toLowerCase()}`;
-    this.props.history.push(url);
+    const { id } = this.props;
+    this.props.history.push(`${config.routes.reports.map}/${id}`);
   };
 
   toggleExpanded = () => {
@@ -149,43 +126,37 @@ class ProjectListItem extends PureComponent {
   };
 
   render() {
-    const {
-      construction_completed: constructionCompleted,
-      photos = [],
-      id,
-      likes,
-      street_name: streetName,
-      title,
-      borough,
-      phase
-    } = this.props;
+    const { id, photo, likes, details, address, status } = this.props;
+    const count = details.number;
+    const subtitle = `${count} ${
+      count === 1 ? 'neuer' : 'neue'
+    } Fahrradbügel gewünscht`;
 
-    const iconSrc = icons[phase];
-    const photo = photos.length ? photos[0] : false;
+    const photoSrc = photo ? photo.src : DefaultPhotoSrc;
+    const photoCopyright = photo ? photo.copyright : '';
+    const statusDisplay = getReportStatusCaption(status);
 
     return (
-      <ItemWrapper onClick={this.toggleExpanded}>
+      <ItemWrapper onClick={this.toggleExpanded} data-testid="report-list-item">
         <ItemContent>
-          <ItemImage src={iconSrc} />
+          <ItemImage />
           <ItemHeader>
-            <ItemTitle>{streetName}</ItemTitle>
-            <Label>
-              {borough} <ProjectLength {...this.props} />
-            </Label>
+            <ItemTitle>{address}</ItemTitle>
+            <Label>Meldung {id}</Label>
           </ItemHeader>
-          <ItemSubTitle>{title}</ItemSubTitle>
+          <ItemSubTitle>{subtitle}</ItemSubTitle>
           <ItemFooter>
             <Likes>
               <HeartIcon />
               <Label>{likes}</Label>
             </Likes>
-            <DateWrapper>Fertigstellung: {constructionCompleted}</DateWrapper>
+            <DateWrapper>Status: {statusDisplay}</DateWrapper>
           </ItemFooter>
         </ItemContent>
         {this.state.isExpanded && (
           <Expansion>
-            <img src={photo.src} alt={title} />
-            <Copyright>{photo.copyright}</Copyright>
+            <img src={photoSrc} alt={subtitle} />
+            <Copyright>{photoCopyright}</Copyright>
             {id && <MapButton onClick={this.onClick}>Zur Karte</MapButton>}
           </Expansion>
         )}
@@ -194,4 +165,4 @@ class ProjectListItem extends PureComponent {
   }
 }
 
-export default ProjectListItem;
+export default ReportListItem;
