+++ src/pages/User/pages/Profile/tabs/ReportLikes.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import Title from '~/components/Title';
 import Text from '~/components/Text';
 import { loadLikes } from '~/pages/User/UserState';
-import ProjectList from '~/components/ProjectList';
+import ProjectList from '~/components2/ProjectList';
 
 const StyledProjectList = styled(ProjectList)`
   margin-top: 1em;
