+++ src/pages/User/pages/Profile/tabs/UserLikes.js
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import Title from '~/components/Title';
 import Text from '~/components/Text';
 import { loadLikes } from '~/pages/User/UserState';
-import ProjectList from '~/components/ProjectList';
+import ProjectList from '~/components2/ProjectList';
 import Select from '~/components/Select';
 
 const StyledSelect = styled(Select)`
